+++ src/ShapeCrawler/Presentations/PresentationCore.cs
@@ -1,10 +1,11 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Validation;
 using ShapeCrawler.Exceptions;
+using A = DocumentFormat.OpenXml.Drawing;
 
 #if NETSTANDARD2_0
 using ShapeCrawler.Extensions;
@@ -81,52 +82,69 @@ internal sealed class PresentationCore
 
     internal void Validate()
     {
-        var nonCritical = new List<ValidationError>
+        var nonCriticalErrorDesc = new List<string>
         {
-            new(
                 "The element has unexpected child element 'http://schemas.openxmlformats.org/drawingml/2006/chart:showDLblsOverMax'.",
-                "/c:chartSpace[1]/c:chart[1]"),
-            new("/c:chartSpace[1]/c:chart[1]/c:extLst[1]/c:ext[1]", "/c:chartSpace[1]/c:chart[1]"),
-            new(
                 "The element has invalid child element 'http://schemas.microsoft.com/office/drawing/2017/03/chart:dataDisplayOptions16'. List of possible elements expected: <http://schemas.microsoft.com/office/drawing/2017/03/chart:dispNaAsBlank>.",
-                "/c:chartSpace[1]/c:chart[1]/c:extLst[1]/c:ext[1]"),
-            new(
                 "The 'uri' attribute is not declared.",
-                "/c:chartSpace[1]/c:chart[1]/c:extLst[1]/c:ext[1]"),
-            new(
                 "The 'mod' attribute is not declared.",
-                "/p:sldLayout[1]/p:extLst[1]"),
-            new(
                 "The 'mod' attribute is not declared.",
-                "/p:sldMaster[1]/p:extLst[1]"),
-            new(
-                "The element has unexpected child element 'http://schemas.openxmlformats.org/drawingml/2006/main:noFill'.",
-                "/p:sld[1]/p:cSld[1]/p:spTree[1]/p:sp[7]/p:spPr[1]")
+                "The element has unexpected child element 'http://schemas.openxmlformats.org/drawingml/2006/main:noFill'."
         };
-
-        var validator = new OpenXmlValidator(FileFormatVersions.Microsoft365);
-        var errors = validator.Validate(this.sdkPresDocument);
-
-        var removing = new List<ValidationErrorInfo>();
-        foreach (var error in errors)
-        {
-            if (nonCritical.Any(x => x.Description == error.Description && x.Path == error.Path?.XPath))
-            {
-                removing.Add(error);
-            }
-        }
-
-        errors = errors.Except(removing);
+        var sdkErrors = new OpenXmlValidator(FileFormatVersions.Microsoft365).Validate(this.sdkPresDocument);
+        sdkErrors = sdkErrors.Where(errorInfo => !nonCriticalErrorDesc.Contains(errorInfo.Description));
 
 #if NETSTANDARD2_0
-        errors = errors.DistinctBy(x => new { x.Description, x.Path?.XPath }).ToList();
+        sdkErrors = sdkErrors.DistinctBy(x => new { x.Description, x.Path?.XPath }).ToList();
 #else
-        errors = errors.DistinctBy(x => new { x.Description, x.Path?.XPath }).ToList();
+        sdkErrors = sdkErrors.DistinctBy(x => new { x.Description, x.Path?.XPath }).ToList();
 #endif
 
+        if (sdkErrors.Any())
+        {
+            throw new SCException("Presentation is invalid.");
+        }
+        
+        var errors = this.ValidateATableRows(this.sdkPresDocument);
         if (errors.Any())
         {
-            throw new SCException("Presentation is invalid. See the Errors property for details.");
+            throw new SCException("Presentation is invalid.");
+        }
+    }
+
+    private IEnumerable<string> ValidateATableRows(PresentationDocument presDocument)
+    {
+        var aTableRows = presDocument.PresentationPart!.SlideParts
+            .SelectMany(slidePart => slidePart.Slide.Descendants<A.TableRow>());
+
+        foreach (var aTableRow in aTableRows)
+        {
+            var aExtLst = aTableRow.GetFirstChild<A.ExtensionList>();
+            if (aExtLst != null)
+            {
+                var lastTableCellIndex = -1;
+                var extListIndex = -1;
+
+                // Find indices of last TableCell and ExtensionList
+                for (int i = 0; i < aTableRow.ChildElements.Count; i++)
+                {
+                    var element = aTableRow.ChildElements[i];
+                    if (element is A.TableCell)
+                    {
+                        lastTableCellIndex = i;
+                    }
+                    else if (element is A.ExtensionList)
+                    {
+                        extListIndex = i;
+                    }
+                }
+
+                // If ExtensionList appears before the last TableCell, yield the error
+                if (extListIndex < lastTableCellIndex)
+                {
+                    yield return "Invalid table row structure: ExtensionList element must appear after all TableCell elements in a TableRow";
+                }
+            }
         }
     }
 }