+++ test/ShapeCrawler.Tests.Unit/ShapeTests.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using ShapeCrawler.Tests.Unit.Helpers;
 using NUnit.Framework;
+using ShapeCrawler.Exceptions;
 
 namespace ShapeCrawler.Tests.Unit;
 
@@ -581,5 +582,198 @@ public class ShapeTests : SCTest
 
         // Assert
         shape.Text.Should().Be("Test");
+    }
+
+    [Test]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 1 Round 0.25", "20.834")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 2 Round 0.25", "20.834")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 3 Round 0.25", "20.834")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 1 Round 0", "0")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 1 Round X", "35")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 1 Round 1", "100")]
+    [SlideShape("057_corner-radius.pptx", 1, "Size 1 Round 0.75", "61.112")]
+    public void CornerSize_getter_returns_values(IShape shape, string expectedSizeStr)
+    {
+        // Arrange
+        var expectedSize = decimal.Parse(expectedSizeStr);
+
+        // Act
+        var actualSize = shape.CornerSize;
+
+        // Assert
+        actualSize.Should().Be(expectedSize);
+    }
+
+    [Test]
+    public void CornerSize_getter_returns_default_values()
+    {
+        // Arrange
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        shapes.AddShape(10, 20, 100, 200, Geometry.RoundedRectangle);
+        var shape = shapes[0];
+
+        // Act
+        var actualSize = shape.CornerSize;
+
+        // Assert
+        actualSize.Should().Be(35m,"Rounded rectangles with no specified corner size behave as if the value was set to 35.");
+    }
+
+    [Test]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded 0.125-ish", "12.29")]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded 0", "0")]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded X", "35")]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded 1", "100")]
+    public void CornerSize_getter_returns_values_for_top_rounded(IShape shape, string expectedSizeStr)
+    {
+        // Arrange
+        var expectedSize = decimal.Parse(expectedSizeStr);
+
+        // Act
+        var actualSize = shape.CornerSize;
+
+        // Assert
+        actualSize.Should().Be(expectedSize);
+    }
+
+    [Test]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded 0.125-ish", "0.5")]
+    [SlideShape("057_corner-radius.pptx", 4, "Top Rounded 0", "0.5")]
+    public void CornerSize_setter_sets_values_for_top_rounded(IShape shape, string expectedSizeStr)
+    {
+        // Arrange
+        var expectedSize = decimal.Parse(expectedSizeStr);
+
+        // Act
+        shape.CornerSize = expectedSize;
+
+        // Assert
+        var actualSize = shape.CornerSize;
+        actualSize.Should().Be(expectedSize);
+    }
+
+    [Test]
+    public void CornerSize_setter_sets_values()
+    {
+        // Arrange
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        shapes.AddShape(10, 20, 100, 200, Geometry.RoundedRectangle);
+        var shape = shapes[0];
+
+        // Act
+        // For most reliable test results, ensure the denominator of the decimal part of
+        // this value is a factor of 50,000. (e.g. here we use 1/10, which will save/load
+        // without loss of precision.)
+        var expected = 0.1m;
+        shape.CornerSize = expected;
+
+        // Assert
+        shape.CornerSize.Should().Be(expected);
+        pres.Validate();
+    }
+
+    [TestCase("RoundedRectangle")]
+    [TestCase("Triangle")]
+    [TestCase("Diamond")]
+    [TestCase("Parallelogram")]
+    [TestCase("Trapezoid")]
+    [TestCase("NonIsoscelesTrapezoid")]
+    [TestCase("DiagonalCornersRoundedRectangle")]
+    [TestCase("TopCornersRoundedRectangle")]
+    [TestCase("SingleCornerRoundedRectangle")]
+    [TestCase("UTurnArrow")]
+    [TestCase("LineInverse")]
+    [TestCase("RightTriangle")]
+    public void Geometry_setter_sets_values(string expectedStr)
+    {
+        // Arrange
+        var expected = (Geometry)Enum.Parse(typeof(Geometry),expectedStr);
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        shapes.AddShape(50, 60, 100, 70);
+        var shape = shapes.Last();
+
+        // Act
+        shape.GeometryType = expected;
+
+        // Assert
+        shape.GeometryType.Should().Be(expected);
+        pres.Validate();
+    }
+
+    [Test]
+    public void Geometry_setter_wont_set_custom()
+    {
+        // Arrange
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        shapes.AddShape(50, 60, 100, 70);
+        var shape = shapes.Last();
+
+        // Act
+        var act = () => shape.GeometryType = Geometry.Custom;
+
+        // Assert
+        act.Should().Throw<SCException>("Custom geometry cannot be set");
+    }
+
+    [Test]
+    public void Geometry_setter_resets_old_adjustments()
+    {
+        // Arrange
+        var pres = new Presentation();
+        var shapes = pres.Slides[0].Shapes;
+        shapes.AddShape(50, 60, 100, 70);
+        var shape = shapes.Last();
+        shape.GeometryType = Geometry.RoundedRectangle;
+        shape.CornerSize = 100;
+
+        // Act
+        shape.GeometryType = Geometry.TopCornersRoundedRectangle;
+
+        // Assert
+        shape.CornerSize.Should().Be(35m,"Default unadjusted corner size is 35");
+    }
+    
+    [Test]
+    public void Name_Setter_sets_shape_name()
+    {
+        // Arrange
+        var pptx = StreamOf("006_1 slides.pptx");
+        var pres = new Presentation(pptx);
+        var stream = new MemoryStream();
+        var shape = pres.Slides[0].Shapes.GetByName("Shape 1");
+
+        // Act
+        shape.Name = "New Name";
+
+        // Assert
+        pres.SaveAs(stream);
+        pres = new Presentation(stream);
+        shape = pres.Slides[0].Shapes.GetByName("New Name");
+        shape.Name.Should().Be("New Name");
+        pres.Validate();
+    }
+    
+    [Test]
+    public void Name_Setter_sets_grouped_shape_name()
+    {
+        // Arrange
+        var pptx = StreamOf("autoshape-grouping.pptx");
+        var pres = new Presentation(pptx);
+        var stream = new MemoryStream();
+        var groupShape = pres.Slides[0].Shapes.GetByName<IGroupShape>("Group 2");
+
+        // Act
+        groupShape.Name = "New Group Name";
+
+        // Assert
+        pres.SaveAs(stream);
+        pres = new Presentation(stream);
+        groupShape = pres.Slides[0].Shapes.GetByName<IGroupShape>("New Group Name");
+        groupShape.Name.Should().Be("New Group Name");
+        pres.Validate();
     }
 }